+++ tests/integration/test_models/standard_model_tests.py
@@ -128,8 +128,9 @@ class OptimisationsTest(object):
 
         self.model = model
 
-    def evaluate_model(self, simplify=False, use_known_evals=False,
-                       to_python=False, to_jax=False):
+    def evaluate_model(
+        self, simplify=False, use_known_evals=False, to_python=False, to_jax=False
+    ):
         result = np.empty((0, 1))
         for eqn in [self.model.concatenated_rhs, self.model.concatenated_algebraic]:
             if simplify: